requirements.txt
# Core web framework
Flask>=3.1.1
flask-sqlalchemy>=3.1.1
flask-cors>=4.0.0
gunicorn>=23.0.0

# Database
SQLAlchemy>=2.0.41
psycopg2-binary>=2.9.10
alembic>=1.13.0

# HTTP requests and utilities
requests>=2.32.4
Werkzeug>=3.1.3
urllib3>=2.0.7

# AI/ML services
google-genai>=1.27.0
openai>=1.97.1

# File processing
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2

# Data validation and processing
email-validator>=2.2.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Authentication and security
cryptography>=41.0.8

# Task scheduling and automation
APScheduler>=3.10.4

# Image processing
Pillow>=10.1.0

# Monitoring and logging
sentry-sdk>=1.38.0

# Additional utilities
python-dateutil>=2.8.2
pytz>=2023.3
click>=8.1.7

# Development tools
python-decouple>=3.8

# Task scheduling
schedule>=1.2.0
